authorization(13-02-2024)

+++ backend/routes/posts.js
@@ -30,7 +30,8 @@ router.post("",checkAuth, multer({ storage: storage }).single("image"), (req, re
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator:req.userData.userId
   });
   post.save().then((result) => {
     res.status(201).json({
@@ -78,8 +79,7 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
-  let imagePath;
-  imagePath=req.body.imagePath;
+  let imagePath =req.body.imagePath;
   if(req.file){
     const url = req.protocol + '://' + req.get("host");
     imagePath = url + "/images/" + req.file.filename
@@ -88,18 +88,29 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath:imagePath
+    imagePath:imagePath,
+    creator:req.userData.userId
   });
   console.log(post);
-  Post.updateOne({ _id: req.body.id }, post).then((result) => {
-    res.status(200).json({ message: "Update successfully!" ,data:post});
+  Post.updateOne({ _id: req.body.id,creator:req.userData.userId }, post).then((result) => {
+    if(result.modifiedCount>0){
+      res.status(200).json({ message: "Update successfully!" ,data:post});
+    }
+    else{
+      res.status(401).json({ message: "Not authorized!"});
+    }
   });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator:req.userData.userId }).then((result) => {
+    if(result.deletedCount>0){
+      res.status(200).json({ message: "Post deleted!" });
+    }
+    else{
+      res.status(401).json({ message: "Not authorized!"});
+    }
+    
   });
 });
